requirements.txt
django==1.11
pillow
pinax-stripe
selenium
celery
mysqlclient
djangorestframework
markdown
django-filter

